fixed file structure for project
