+++ GenericsHomework.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34408.163
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GenericsHomework", "GenericsHomework\GenericsHomework.csproj", "{D158A4E7-D8CF-4849-9960-7D5D8E18851C}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GenericsTests", "GenericsTests\GenericsTests.csproj", "{E1A71020-8CA9-41E7-823A-B1781E7100D7}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{1EC52FA5-12F0-4B4B-B3C6-E9CE1F6A8FF0}"
+	ProjectSection(SolutionItems) = preProject
+		Directory.Build.props = Directory.Build.props
+		Directory.Build.targets = Directory.Build.targets
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D158A4E7-D8CF-4849-9960-7D5D8E18851C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D158A4E7-D8CF-4849-9960-7D5D8E18851C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D158A4E7-D8CF-4849-9960-7D5D8E18851C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D158A4E7-D8CF-4849-9960-7D5D8E18851C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E1A71020-8CA9-41E7-823A-B1781E7100D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E1A71020-8CA9-41E7-823A-B1781E7100D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E1A71020-8CA9-41E7-823A-B1781E7100D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E1A71020-8CA9-41E7-823A-B1781E7100D7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D4AEFBCF-5440-49A8-9280-14523DBDD1D8}
+	EndGlobalSection
+EndGlobal
