--- GenericsHomework/GenericsHomework.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.8.34408.163
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenericsHomework", "GenericsHomework\GenericsHomework.csproj", "{D158A4E7-D8CF-4849-9960-7D5D8E18851C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GenericsTests", "GenericsTests\GenericsTests.csproj", "{E1A71020-8CA9-41E7-823A-B1781E7100D7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D158A4E7-D8CF-4849-9960-7D5D8E18851C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D158A4E7-D8CF-4849-9960-7D5D8E18851C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D158A4E7-D8CF-4849-9960-7D5D8E18851C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D158A4E7-D8CF-4849-9960-7D5D8E18851C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E1A71020-8CA9-41E7-823A-B1781E7100D7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1A71020-8CA9-41E7-823A-B1781E7100D7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E1A71020-8CA9-41E7-823A-B1781E7100D7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E1A71020-8CA9-41E7-823A-B1781E7100D7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D4AEFBCF-5440-49A8-9280-14523DBDD1D8}
-	EndGlobalSection
-EndGlobal
