--- GenericsHomework/GenericsHomework/Directory.Build.props
@@ -1,15 +1,0 @@
-<Project>
-  <!-- See https://aka.ms/dotnet/msbuild/customize for more details on customizing your build -->
-  <PropertyGroup>
-	  <Deterministic>true</Deterministic>
-	  <TargetFramework>net8.0</TargetFramework>
-	  <LangVersion>12.0</LangVersion>
-	  <ImplicitUsings>enable</ImplicitUsings>
-	  <Nullable>enable</Nullable>
-	  <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-	  <EnableNETAnalyzers>true</EnableNETAnalyzers>
-	  <CodeAnalysisTreatWarningsAsErrors>true</CodeAnalysisTreatWarningsAsErrors>
-	  <EnforceCodeStyleInBuild>true</EnforceCodeStyleInBuild>
-
-  </PropertyGroup>
-</Project>
